[FE-149] fix: useRef로 뒤로가기 시 안되던 스크롤 버그 픽스
