+++ src/types/replyData.ts
@@ -8,6 +8,7 @@ export interface CommentData {
   writer?: string
   recordwriter?: string
   recordId?: string | undefined
+  isScroll?: boolean
 }
 
 export interface CommentRequestDtoType {
